cerebus/requirements.txt
# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1

# Redis
redis==5.2.0

# ML / Guardrail
transformers==4.45.2
torch==2.4.1
peft==0.13.0
datasets==3.0.1
detoxify==0.5.2
sentence-transformers==3.1.1
faiss-cpu==1.9.0

# NLP utilities
spacy==3.8.2
evaluate==0.4.3

# LangChain (RAG)
langchain==0.3.3
langchain-community==0.3.2
langchain-huggingface==0.1.2
